ocr-service/requirements.txt
# Project: Vitamend
# Creator: Rachit Kumar Tiwari
fastapi==0.115.0
uvicorn==0.30.1
motor==3.5.1
python-dotenv==1.0.1
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
numpy==1.26.4
pyzbar==0.1.9
python-dateutil==2.9.0.post0
pytest==8.3.2
